第6次上传,商品表分页查询中，模糊查询商品名. 添加单个或多个条件模糊查询，价格范围。 通过外连接得到的数据用Map返回时需要在sql中重命名， 防止重名数据缺失列如u.name与g.name重名. 修改Amount类型integer-->double,有散装称重商品

+++ src/main/java/com/example/demo/domain/InstockDetail.java
@@ -9,14 +9,14 @@ public class InstockDetail {
     private Integer id;
     private Integer goodsId;//商品id
     private double originalPrice;//商品原价
-    private Integer amount;//总计
+    private double amount;//总计
     private Integer instockId;//库存id
     private Integer status;//状态
 
     public InstockDetail() {
     }
 
-    public InstockDetail(Integer id, Integer goodsId, double originalPrice, Integer amount, Integer instockId, Integer status) {
+    public InstockDetail(Integer id, Integer goodsId, double originalPrice, double amount, Integer instockId, Integer status) {
         this.id = id;
         this.goodsId = goodsId;
         this.originalPrice = originalPrice;
@@ -49,11 +49,11 @@ public class InstockDetail {
         this.originalPrice = originalPrice;
     }
 
-    public Integer getAmount() {
+    public double getAmount() {
         return amount;
     }
 
-    public void setAmount(Integer amount) {
+    public void setAmount(double amount) {
         this.amount = amount;
     }
